Undoing Changes on File

+++ app/controllers/wikis_controller.rb
@@ -1,7 +1,14 @@
 class WikisController < ApplicationController
 
+
   def index
     @wikis = Wiki.all
+    authorize @wikis
+    # @wikis = Wiki.visible_to(current_user)
+    #
+    # if current_user.premium? || current_user.admin?
+    #   @wikis = Wiki.all
+    # end
   end
 
   def show
@@ -9,43 +16,63 @@ class WikisController < ApplicationController
   end
 
   def new
+    @user = current_user
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
-    @wiki = Wiki.new
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
+    @wiki = Wiki.new(wiki_params)
+    @wiki.user = current_user
+    authorize @wiki
 
-    if @wiki.save
-      flash[:notice] = "Wiki was saved."
-      redirect_to @wiki
-    else
-      flash.now[:alert] = "There was an error saving the wiki. Please try again."
+    # @wiki.title = params[:wiki][:title]
+    # @wiki.body = params[:wiki][:body]
+
+    if @wiki.private && @wiki.user.standard?
+      flash[:notice] = "You need a premium account to make your Wiki private."
       render :new
+    else
+      if @wiki.save
+        flash[:notice] = "Wiki was saved."
+        redirect_to @wiki
+      else
+        flash.now[:alert] = "There was an error saving the wiki. Please try again."
+        render :new
+      end
     end
   end
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
+    authorize @wiki
+    @wiki.assign_attributes(wiki_params)
 
-    if @wiki.save
-      flash[:notice] = "wiki was updated."
-      redirect_to @wiki
-    else
-      flash.now[:alert] = "There was an error saving the wiki. Please try again."
+    # @wiki.title = params[:wiki][:title]
+    # @wiki.body = params[:wiki][:body]
+
+    if @wiki.private && @wiki.user.standard?
+      flash[:alert] = "You need a Premium account to make your Wiki private."
       render :edit
+    else
+      if @wiki.save
+        flash[:notice] = "wiki was updated."
+        redirect_to @wiki
+      else
+        flash.now[:alert] = "There was an error saving the wiki. Please try again."
+        render :edit
+      end
     end
   end
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
@@ -57,6 +84,10 @@ class WikisController < ApplicationController
   end
 
   private
+
+  def wiki_params
+    params.require(:wiki).permit(:title, :body, :private)
+  end
 
   def authorize_user
     unless current_user.admin? || current_user.premium?
